refactor(renderer): Rename archivefile to af
