+++ cmd/dukkha/addon/renderers.go
@@ -2,7 +2,7 @@ package addon
 
 import (
 	// Add default disabled renderers
-	_ "arhat.dev/dukkha/pkg/renderer/archivefile"
+	_ "arhat.dev/dukkha/pkg/renderer/af"
 	_ "arhat.dev/dukkha/pkg/renderer/git"
 	_ "arhat.dev/dukkha/pkg/renderer/http"
 	_ "arhat.dev/dukkha/pkg/renderer/input"
