+++ pkg/renderer/af/fixtures/002-object-spec.yaml
