+++ pkg/renderer/af/rar.go
@@ -1,4 +1,4 @@
-package archivefile
+package af
 
 import (
 	"fmt"
