+++ pkg/renderer/af/tar_test.go
@@ -1,4 +1,4 @@
-package archivefile
+package af
 
 import (
 	"archive/tar"
