+++ pkg/renderer/af/testdata/004.tar.lzma
