+++ pkg/renderer/af/testdata/_archive_content/level-1/level-1-data-symlink
