+++ pkg/renderer/af/testdata/_archive_content/level-1/level-2/top-level-data-symlink
