+++ pkg/renderer/af/testdata/gen.sh
