+++ pkg/tools/archive/fixtures/create/001-tar.yaml
@@ -10,9 +10,9 @@ task:
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.tar" -}}
 ---
 actual:
-  top-level@template?str|archivefile?str: |-
+  top-level@template?str|af?str: |-
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.tar" -}}:top-level.txt
-  root-top-level@template?str|archivefile?str: |-
+  root-top-level@template?str|af?str: |-
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.tar" -}}:/top-level.txt
 expected:
   top-level@file?str: testdata/top-level.txt
