+++ pkg/tools/archive/fixtures/create/006-tar-deflate.yaml
@@ -20,9 +20,9 @@ task:
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.tar.df" -}}
 ---
 actual:
-  top-level@template?str|archivefile?str: |-
+  top-level@template?str|af?str: |-
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.tar.df" -}}:top-level.txt
-  root-top-level@template?str|archivefile?str: |-
+  root-top-level@template?str|af?str: |-
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.tar.df" -}}:/top-level.txt
 expected:
   top-level@file?str: testdata/top-level.txt
