+++ pkg/tools/archive/fixtures/create/103-zip-bzip2.yaml
@@ -20,9 +20,9 @@ task:
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.zip" -}}
 ---
 actual:
-  top-level@template?str|archivefile?str: |-
+  top-level@template?str|af?str: |-
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.zip" -}}:top-level.txt
-  root-top-level@template?str|archivefile?str: |-
+  root-top-level@template?str|af?str: |-
     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.zip" -}}:/top-level.txt
 expected:
   top-level@file?str: testdata/top-level.txt
