+++ pkg/tools/archive/fixtures/create/106-zip-gz-error.yaml
@@ -21,9 +21,9 @@ task:
 ---
 expect_err: true
 # actual:
-#   top-level@template?str|archivefile?str: |-
+#   top-level@template?str|af?str: |-
 #     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.zip" -}}:top-level.txt
-#   root-top-level@template?str|archivefile?str: |-
+#   root-top-level@template?str|af?str: |-
 #     {{- filepath.Join env.DUKKHA_CACHE_DIR "test.zip" -}}:/top-level.txt
 # expected:
 #   top-level@file?str: testdata/top-level.txt
