+++ sonar-project.properties
@@ -1,6 +1,6 @@
 sonar.projectKey=arhat-dev_dukkha
 
 sonar.go.coverage.reportPaths=build/test-profile/coverage*.txt
-sonar.go.tests.reportPaths=build/test-profile/result-pkg-cache.json,build/test-profile/result-pkg-cmd-debug.json,build/test-profile/result-pkg-cmd-diff.json,build/test-profile/result-pkg-cmd-render.json,build/test-profile/result-pkg-cmd-utils.json,build/test-profile/result-pkg-cmd.json,build/test-profile/result-pkg-conf.json,build/test-profile/result-pkg-constant.json,build/test-profile/result-pkg-diff.json,build/test-profile/result-pkg-dukkha.json,build/test-profile/result-pkg-matrix.json,build/test-profile/result-pkg-renderer-archivefile.json,build/test-profile/result-pkg-renderer-echo.json,build/test-profile/result-pkg-renderer-env.json,build/test-profile/result-pkg-renderer-file.json,build/test-profile/result-pkg-renderer-git.json,build/test-profile/result-pkg-renderer-http.json,build/test-profile/result-pkg-renderer-shell.json,build/test-profile/result-pkg-renderer-template.json,build/test-profile/result-pkg-renderer-transform.json,build/test-profile/result-pkg-renderer.json,build/test-profile/result-pkg-templateutils.json,build/test-profile/result-pkg-tools-buildah.json,build/test-profile/result-pkg-tools-git.json,build/test-profile/result-pkg-tools-github.json,build/test-profile/result-pkg-tools-golang.json,build/test-profile/result-pkg-tools.json,build/test-profile/result-pkg-utils.json
+sonar.go.tests.reportPaths=build/test-profile/result-pkg-cache.json,build/test-profile/result-pkg-cmd-debug.json,build/test-profile/result-pkg-cmd-diff.json,build/test-profile/result-pkg-cmd-render.json,build/test-profile/result-pkg-cmd-utils.json,build/test-profile/result-pkg-cmd.json,build/test-profile/result-pkg-conf.json,build/test-profile/result-pkg-constant.json,build/test-profile/result-pkg-diff.json,build/test-profile/result-pkg-dukkha.json,build/test-profile/result-pkg-matrix.json,build/test-profile/result-pkg-renderer-af.json,build/test-profile/result-pkg-renderer-echo.json,build/test-profile/result-pkg-renderer-env.json,build/test-profile/result-pkg-renderer-file.json,build/test-profile/result-pkg-renderer-git.json,build/test-profile/result-pkg-renderer-http.json,build/test-profile/result-pkg-renderer-shell.json,build/test-profile/result-pkg-renderer-template.json,build/test-profile/result-pkg-renderer-transform.json,build/test-profile/result-pkg-renderer.json,build/test-profile/result-pkg-templateutils.json,build/test-profile/result-pkg-tools-buildah.json,build/test-profile/result-pkg-tools-git.json,build/test-profile/result-pkg-tools-github.json,build/test-profile/result-pkg-tools-golang.json,build/test-profile/result-pkg-tools.json,build/test-profile/result-pkg-utils.json
 
 sonar.go.exclusions=vendor/**,**/vendor/**,third_party/**,**/third_party/**,**/*_test.go
